requirements.txt
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
scipy>=1.7.0
openpyxl>=3.0.0
transformers>=4.30.0
torch>=2.0.0
Pillow>=9.0.0 